Fix structure for metrics/tracing@
