+++ PFM.Website.Reboot/src/PFM.Services/Configurations/AuthHeaderHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net.Http.Headers;
+using Microsoft.AspNetCore.Http;
 
 namespace PFM.Website.Configurations
 {
