+++ PFM.Website.Reboot/src/PFM.Services/Configurations/PfmApiOptions.cs
@@ -1,0 +1,7 @@
+namespace PFM.Services.Configurations;
+
+public class PfmApiOptions
+{
+    public bool Enabled { get; init; } = false;
+    public string EndpointUrl { get; init; } = string.Empty;
+}
