+++ PFM.Website.Reboot/src/PFM.Services/Monitoring/Metrics/MetricsOptions.cs
@@ -3,4 +3,6 @@ namespace PFM.Services.Monitoring.Metrics;
 public class MetricsOptions
 {
     public bool Debug { get; init; } = false;
+    
+    public string ServiceName { get; init; } = string.Empty;
 }
