+++ PFM.Website.Reboot/src/PFM.Services/Monitoring/Tracing/TracingOptions.cs
@@ -3,4 +3,6 @@ namespace PFM.Services.Monitoring.Tracing;
 public class TracingOptions
 {
     public bool Debug { get; init; } = false;
+    
+    public string ServiceName { get; init; } = string.Empty;
 }
