+++ PFM.Website.Reboot/src/PFM.Website/Configurations/ServiceCollectionExtensions.cs
@@ -65,90 +65,6 @@ namespace PFM.Website.Configurations
 
             return services;
 		}
-        
-        public static IServiceCollection AddPfmApi(this IServiceCollection services, IConfiguration configuration, ApplicationSettings appSettings, bool isDevelopmentEnvironment)
-        {
-            if (!appSettings.PfmApiOptions.Enabled)
-            {
-                services
-                    .AddSingleton<IExpenseTypeApi, ExpenseTypeInMemory>()
-                    .AddSingleton<IBankApi, BankInMemory>()
-                    .AddSingleton<ICountryApi, CountryInMemory>()
-                    .AddSingleton<IBankAccountApi, BankAccountInMemory>()
-                    .AddSingleton<ICurrencyApi, CurrencyInMemory>()
-                    .AddSingleton<IIncomeApi, IncomeInMemory>()
-                    .AddSingleton<ISavingApi, SavingInMemory>()
-                    .AddSingleton<IAtmWithdrawApi, AtmWithdrawInMemory>()
-                    .AddSingleton<IExpenseApi, ExpenseInMemory>()
-                    .AddSingleton<IPaymentMethodApi, PaymentMethodInMemory>()
-                    .AddSingleton<IMovementSummaryApi, MovementSummaryInMemory>()
-                    .AddSingleton<IBudgetPlanApi, BudgetPlanInMemory>();
-                return services;
-            }
-
-            var refitSettings = new RefitSettings()
-            {
-                ContentSerializer = new SystemTextJsonContentSerializer(
-                    new JsonSerializerOptions()
-                    {
-                        PropertyNameCaseInsensitive = true,
-                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                        Converters =
-                        {
-                            new ObjectToInferredTypesConverter(),
-                        }
-                    }
-                ),
-                ExceptionFactory = httpResponse =>
-                {
-                    return Task.FromResult<Exception?>(null);
-                }
-            };
-
-            var httpClientHandler = !isDevelopmentEnvironment ? new HttpClientHandler() : new HttpClientHandler { ServerCertificateCustomValidationCallback = (message, cert, chain, sslErrors) => true };
-
-            var apiConfigs = appSettings.PfmApiOptions.EndpointUrl;
-            services
-                .AddPfmApiClient<IExpenseTypeApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<ICountryApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IBankApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<ICurrencyApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IBankAccountApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IIncomeApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<ISavingApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IAtmWithdrawApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IExpenseApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IPaymentMethodApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IMovementSummaryApi>(apiConfigs, refitSettings, httpClientHandler)
-                .AddPfmApiClient<IBudgetPlanApi>(apiConfigs, refitSettings, httpClientHandler);
-
-            return services;
-        }
-
-        private static IServiceCollection AddPfmApiClient<T>(this IServiceCollection services, string baseUrl, RefitSettings refitSettings, HttpClientHandler httpClientHandler)
-            where T : class
-        {
-            services
-               .AddRefitClient<T>(refitSettings)
-               .ConfigureHttpClient(c => c.BaseAddress = new Uri(baseUrl))
-               .ConfigurePrimaryHttpMessageHandler(() => httpClientHandler)
-               .AddHttpMessageHandler<AuthHeaderHandler>();
-
-            return services;
-        }
-
-        public static IServiceCollection AddObjectMapper(this IServiceCollection services)
-        {
-            var configuration = new MapperConfiguration(cfg => {
-                cfg.AddProfile<ModelToResponseProfile>();
-                cfg.AddProfile<RequestToModelProfile>();
-            });
-
-            services.AddSingleton(configuration);
-            services.AddSingleton<IMapper>(configuration.CreateMapper());
-
-            return services;
-        }
     }
 }
