+++ PFM.Website.Reboot/src/PFM.Website/Monitoring/Metrics/ServiceCollectionExtensions.cs
@@ -1,0 +1,20 @@
+using OpenTelemetry.Metrics;
+using PFM.Services.Monitoring.Metrics;
+
+namespace PFM.Website.Monitoring.Metrics;
+
+public static class ServiceCollectionExtensions
+{
+    public static IServiceCollection ConfigureWebsiteMetrics(this IServiceCollection services, MetricsOptions options)
+    { 
+        services.AddSingleton<IDashboardMetrics, DashboardMetrics>();
+        services.ConfigureMetrics(options);
+        
+        services.ConfigureOpenTelemetryMeterProvider(builder =>
+        {
+            builder.AddMeter(DashboardMetrics.MeterName);
+        });
+
+        return services;
+    }
+}
