+++ PFM.Website.Reboot/src/PFM.Website/Pages/Index.razor
@@ -2,8 +2,8 @@
 @page "/{accountId:int}"
 @using PFM.Models
 @using PFM.Services
-@using PFM.Services.Monitoring.Metrics
 @using PFM.Website.Components.Dashboard
+@using PFM.Website.Monitoring.Metrics
 
 @inject BankAccountService BankAccountService
 @inject NavigationManager NavigationManager
