+++ PFM.Website.Reboot/src/PFM.Website/appsettings.json
@@ -39,10 +39,12 @@
       "EndpointUrl": "https://localhost:4431/"
     },
     "MetricsOptions": {
-      "Debug": false
+      "Debug": false,
+      "ServiceName": "PFM.Website"
     },
     "TracingOptions": {
-      "Debug": false
+      "Debug": false,
+      "ServiceName": "PFM.Website"
     }
   }
 }
